Module common_helpers::avro

Enums

AvroError

Represents error types returned by the avro module.

Functions

fingerprint_raw_schema

Function to convert a raw schema into serialized Avro schema. If schema is malformed or invalid, returns an error.

fingerprint_raw_schema_list

Function to convert a list of raw schema into serialized Avro schema. If schema is malformed or invalid, it is set to Null.

get_schema_data_map

Function to get serialized datum data for a given schema into hashmap.

get_schema_data_writer

Function to get the schema writer with default container as Vec

populate_schema_and_serialize

Function to populate a given schema with data and return serialized record.

translate_schema

Function to convert a serialized Avro schema into Avro Schema type. If schema is malformed or invalid, returns an error.

translate_schemas

Function to convert a list of serialized Avro schema into Avro Schema type. If schema is malformed or invalid, it is set to Null.

validate_raw_avro_schema

Function to validate incoming json serialized schema against avro schema.