Function common_helpers::avro::fingerprint_raw_schema_list

pub fn fingerprint_raw_schema_list(
    raw_schema: &[&str]
) -> Result<(Vec<Schema>, Vec<Vec<u8>>), AvroError>

Function to convert a list of raw schema into serialized Avro schema. If schema is malformed or invalid, it is set to Null.

Arguments

raw_schema - raw schema list to be converted

Returns

Result<(Vec<Schema>, Vec<Vec<u8>>), AvroError> - structured and serialized Avro schemas

Examples

use common_helpers::avro;
use common_helpers::types::*;
let raw_schema = r#"{"type": "record", "name": "User", "fields": [{"name": "name", "type": "string"}, {"name": "favorite_number", "type": "int"}]}"#;
let vec_raw_schema: [&str; 1] = [raw_schema];
let schema_result = avro::fingerprint_raw_schema_list(&vec_raw_schema);
assert!(schema_result.is_ok());
let serialized_schemas = schema_result.unwrap().1;